requirements.txt
prefect==2.14.20
griffe==0.49.0
tensorflow
Pillow

